[WFLY-12730]:Move SimpleGroup to a public outter class
